+++ test/mocks/jsapi3x.js
@@ -1,3 +1,3 @@
-// this is defined in spec
 console.log('3.x');
+// NOTE: this is defined in /src/script.test.ts
 stubRequire();