+++ CHANGELOG.md
@@ -4,7 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- `loadScript()` takes a new `version` option to load a specific version from the CDN
+- passing `css: true` to `loadScript()` will load the styles for the CDN version
+- `loadCss()` defaults to loading the latest 4.x styles if no arguments are passed
+- `loadCss()` can take a version as a string to load a version's styles from the CDN
 ### Changed
+- split source code into modules
+- tests are now written in TypeScript and loaded via karma-typescript
+- updated to recent versions of TypeScript, Karma, & Jasmine
 ### Fixed
 ### Removed
 ### Breaking